Split traits out
